fix: rehash strong exception guarantee + comprehensive test infrastructure
Core fix:
- hash_table5.hpp: reorder rehash() to allocate new memory BEFORE updating
  member variables, ensuring strong exception guarantee on OOM
- Remove incorrect noexcept from alloc_bucket/rehash/reserve/check_expand_need
  so bad_alloc can propagate instead of calling std::terminate

New tests:
- test_complex_conditions.cpp: 14 tests covering concurrent reads, burst
  insert/erase, high load factor, OOM recovery, 10M sustained ops, performance
  metrics, and adversarial inputs (2.47M assertions)
- test_emilib_lifecycle.cpp: 19 tests for emilib clone/rehash/clear/move/erase
  lifecycle (12.9K assertions)

CI improvements:
- fuzz.yml: weekly scheduled libfuzzer job
- ci.yml: MSan/coverage expanded, benchmark threshold fixed (10%->20%),
  test_complex_conditions added to build/windows-msvc/coverage jobs
- test-list.sh: SSOT for CI test file lists

Engineering:
- CMakeLists.txt: emhash_add_bench() helper reduces 30 lines of boilerplate
- bench/research/: 12 debug/diag scripts moved from bench/
- bench/README.md: documents official vs research benchmarks
- Doxyfile + docs-build CI job for auto-generated API docs
- ADR 005/006/007: prefetch disable, bitmask layout, swiss table experiment
- CHANGELOG: fix threshold references (25%->20%), remove non-existent file refs
- clang-format: fix emihmap2/3/4, emihset3, ebench format issues
- emihmap4.hpp: replace __int64 with uint64_t (fixes Clang/Windows -Werror)

+++ tests/unit/test_stress_correctness.cpp
@@ -734,7 +734,7 @@ TEST_CASE_TEMPLATE("random string key insert/erase stress", Map, AllStringMaps) 
 TEST_CASE_TEMPLATE("reserve then fill to boundary", Map, AllIntMaps) {
     Map m;
     m.reserve(1000);
-    auto bc = m.bucket_count();
+    auto bc = static_cast<size_t>(m.bucket_count());
     // Fill exactly to capacity
     for (size_t i = 0; i < bc; i++)
         m[static_cast<int>(i)] = static_cast<int>(i);
